implementation eines SessionsManagers. NICHT FINAL!
